refactor the duplicated code in otlp exporter with transform part
