+++ exporters/otlp/internal/transform/attribute.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrictransform // import "go.opentelemetry.io/otel/exporters/otlp/otlpmetric/internal/metrictransform"
+package transform // import "go.opentelemetry.io/otel/exporters/otlp/transform"
 
 import (
 	"go.opentelemetry.io/otel/attribute"
