+++ exporters/otlp/internal/transform/resource.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tracetransform // import "go.opentelemetry.io/otel/exporters/otlp/otlptrace/internal/tracetransform"
+package transform // import "go.opentelemetry.io/otel/exporters/otlp/transform"
 
 import (
 	"go.opentelemetry.io/otel/sdk/resource"
