+++ exporters/otlp/otlptrace/internal/tracetransform/span_test.go
@@ -25,6 +25,7 @@ import (
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
+	"go.opentelemetry.io/otel/exporters/otlp/internal/transform"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/resource"
 	tracesdk "go.opentelemetry.io/otel/sdk/trace"
@@ -98,7 +99,11 @@ func TestSpanEvent(t *testing.T) {
 	eventTimestamp := uint64(1589932800 * 1e9)
 	assert.Equal(t, &tracepb.Span_Event{Name: "test 1", Attributes: nil, TimeUnixNano: eventTimestamp}, got[0])
 	// Do not test Attributes directly, just that the return value goes to the correct field.
-	assert.Equal(t, &tracepb.Span_Event{Name: "test 2", Attributes: KeyValues(attrs), TimeUnixNano: eventTimestamp, DroppedAttributesCount: 2}, got[1])
+	assert.Equal(t, &tracepb.Span_Event{Name: "test 2",
+		Attributes:             transform.KeyValues(attrs),
+		TimeUnixNano:           eventTimestamp,
+		DroppedAttributesCount: 2,
+	}, got[1])
 }
 
 func TestNilLinks(t *testing.T) {
@@ -137,7 +142,7 @@ func TestLinks(t *testing.T) {
 	assert.Equal(t, expected, got[0])
 
 	// Do not test Attributes directly, just that the return value goes to the correct field.
-	expected.Attributes = KeyValues(attrs)
+	expected.Attributes = transform.KeyValues(attrs)
 	assert.Equal(t, expected, got[1])
 
 	// Changes to our links should not change the produced links.
@@ -287,7 +292,7 @@ func TestSpanData(t *testing.T) {
 		Status:                 status(spanData.Status.Code, spanData.Status.Description),
 		Events:                 spanEvents(spanData.Events),
 		Links:                  links(spanData.Links),
-		Attributes:             KeyValues(spanData.Attributes),
+		Attributes:             transform.KeyValues(spanData.Attributes),
 		DroppedAttributesCount: 1,
 		DroppedEventsCount:     2,
 		DroppedLinksCount:      3,
@@ -296,7 +301,7 @@ func TestSpanData(t *testing.T) {
 	got := Spans(tracetest.SpanStubs{spanData}.Snapshots())
 	require.Len(t, got, 1)
 
-	assert.Equal(t, got[0].GetResource(), Resource(spanData.Resource))
+	assert.Equal(t, got[0].GetResource(), transform.Resource(spanData.Resource))
 	assert.Equal(t, got[0].SchemaUrl, spanData.Resource.SchemaURL())
 	ilSpans := got[0].GetInstrumentationLibrarySpans()
 	require.Len(t, ilSpans, 1)
