+++ go.mod
@@ -7,7 +7,9 @@ require (
 	github.com/go-logr/stdr v1.2.2
 	github.com/google/go-cmp v0.5.7
 	github.com/stretchr/testify v1.7.0
+	go.opentelemetry.io/otel/sdk v0.0.0-00010101000000-000000000000
 	go.opentelemetry.io/otel/trace v1.5.0
+	go.opentelemetry.io/proto/otlp v0.12.0
 )
 
 replace go.opentelemetry.io/otel => ./
